+++ Assets/Scenes/StaminaParry.unity.meta
