+++ Assets/Scripts/PlayerCombat.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 78e8022d4cd4301438ff8602873185be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
