+++ Assets/Scripts/PlayerStamina.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 91674db431276e446821b05882eee380
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
